movies/requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==20.3.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.11.1
bitarray==2.6.0
black==22.6.0
bleach==5.0.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
cmake==3.22.4
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cycler==0.11.0
cytoolz==0.12.0
debugpy==1.6.2
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.1
django-crispy-forms==1.14.0
django-filter==23.4
django-mathfilters==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
entrypoints==0.4
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==0.9.1
fastjsonschema==2.16.1
fonttools==4.37.0
frozenlist==1.3.1
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
hexbytes==0.3.0
httplib2==0.20.4
idna==3.3
importlib-metadata==6.8.0
inflection==0.5.1
ipfshttpclient==0.8.0a2
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itypes==1.2.0
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
lib50==3.0.4
lru-dict==1.1.8
Markdown==3.5.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netaddr==0.8.0
notebook==6.4.12
numpy==1.23.4
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prettytable==3.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.5
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pygame==2.1.2
Pygments==2.12.0
pygsheets==2.0.5
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pywin32==304
pywinpty==2.0.7
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rlp==2.0.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
submit50==3.1.1
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
turtle==0.0.1
typing_extensions==4.3.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
varint==1.0.2
wcwidth==0.2.5
web3==5.31.1
webencodings==0.5.1
websockets==9.1
whitenoise==6.2.0
widgetsnbextension==3.6.1
yarl==1.8.1
zipp==3.17.0
